django_project/requirements.txt
beautifulsoup4==4.7.1
bs4==0.0.1
certifi==2019.3.9
chardet==3.0.4
Django==2.1.7
django-cors-headers==2.5.2
django-crispy-forms==1.7.2
idna==2.8
lxml==4.3.3
Pillow==5.4.1
pkg-resources==0.0.0
pytz==2018.9
requests==2.21.0
rerequests==1.0.0b0
soupsieve==1.9
urllib3==1.24.1
